debug

+++ datasets/usps.py
@@ -66,7 +66,7 @@ class USPS(data.Dataset):
         img, label = self.train_data[index], self.train_labels[index]
         if self.transform is not None:
             img = self.transform(img)
-        return img, label
+        return img, label.astype("int64")
 
     def __len__(self):
         """Return size of dataset."""
